move the Executor to designated file
